Implements basic AoC setup and solver runner
Adds basic project structure for Advent of Code solutions.
Includes utility functions for running solvers, reading inputs,
and updating metadata.

Sets up a command-line interface for specifying the year, day,
and part of the challenge to run.

Adds initial files for 2024 and day 1 part 1.
Includes .gitignore updates and ruff as a dev dependency.