Adds solutions for Day 8, parts 1 and 2
Implements solutions for Day 8, parts 1 and 2 of the Advent of Code 2025 challenge.

Part 1 calculates the product of the sizes of the three largest circuits formed by connecting junction boxes based on shortest distances.

Part 2 connects junction boxes in order of increasing distance until all boxes are in one circuit and returns the product of the X coordinates of the final pair that completes the circuit.